move batch2 to datalad_next.runner
